Рефакторинг

+++ src/main/java/ru/demo/homework/entity/ProductEntity.java
@@ -3,12 +3,14 @@ package ru.demo.homework.entity;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 @Data
 @Accessors(chain = true)
 public class ProductEntity {
     private Long id;
     private String account;
-    private Float balance;
+    private BigDecimal balance;
     private Integer type;
     private Long userId;
 }
